rag-agent/requirements.txt
# Core RAG pipeline
llama-index-core>=0.12.0
llama-index-embeddings-google-genai>=0.4.0
llama-index-llms-google-genai>=0.4.0
llama-index-vector-stores-supabase>=0.3.0

# Direct DB access for hybrid search RPC
asyncpg>=0.29.0

# Environment
python-dotenv>=1.0.0

# Cross-encoder reranking (Phase 4)
flashrank>=0.2.0

# MCP server (Phase 4)
mcp>=1.0.0

# Helper tools (rag-agent/tools/)
google-generativeai>=0.8.0
google-genai>=1.0.0
